cpufreq: Register governors at core_initcall
Currently, most CPUFreq governors are registered at the core_initcall
time when the given governor is the default one, and the module_init
time otherwise.

In preparation for letting users specify the default governor on the
kernel command line, change all of them to be registered at the
core_initcall unconditionally, as it is already the case for the
schedutil and performance governors. This will allow us to assume
that builtin governors have been registered before the built-in
CPUFreq drivers probe.

And since all governors have similar init/exit patterns now, introduce
two new macros, cpufreq_governor_{init,exit}(), to factorize the code.

Acked-by: Viresh Kumar <viresh.kumar@linaro.org>
[ rjw: Changelog ]

Signed-off-by: TogoFire <togofire@mailfence.com>
Change-Id: I0b0085dbcc15375af745030ea79b5e060fa13550
Signed-off-by: prabhatKrMishra <mprabhat774@gmail.com>
Signed-off-by: 0xA11CE613 <mprabhat774@gmail.com>